+++ controllers/keycloakrealmrole/terminator.go
@@ -10,12 +10,18 @@ import (
 )
 
 type terminator struct {
-	realmName, realmRoleName string
-	kClient                  keycloak.Client
+	realmName, realmRoleName    string
+	kClient                     keycloak.Client
+	preserveResourcesOnDeletion bool
 }
 
 func (t *terminator) DeleteResource(ctx context.Context) error {
 	log := ctrl.LoggerFrom(ctx)
+	if t.preserveResourcesOnDeletion {
+		log.Info("PreserveResourcesOnDeletion is enabled, skipping deletion.")
+		return nil
+	}
+
 	log.Info("Start deleting keycloak realm role")
 
 	if err := t.kClient.DeleteRealmRole(ctx, t.realmName, t.realmRoleName); err != nil {
@@ -27,10 +33,11 @@ func (t *terminator) DeleteResource(ctx context.Context) error {
 	return nil
 }
 
-func makeTerminator(realmName, realmRoleName string, kClient keycloak.Client) *terminator {
+func makeTerminator(realmName, realmRoleName string, kClient keycloak.Client, preserveResourcesOnDeletion bool) *terminator {
 	return &terminator{
-		realmRoleName: realmRoleName,
-		realmName:     realmName,
-		kClient:       kClient,
+		realmRoleName:               realmRoleName,
+		realmName:                   realmName,
+		kClient:                     kClient,
+		preserveResourcesOnDeletion: preserveResourcesOnDeletion,
 	}
 }
